Move common parts to one place
