+++ s3torchbenchmarking/src/s3torchbenchmarking/dcp_fsdp/benchmark.py
@@ -4,18 +4,12 @@
 import logging
 import functools
 from multiprocessing.queues import Queue
-from pathlib import Path
 from time import perf_counter
-from typing import List, Tuple
+from typing import Tuple
 
 import hydra
-import pandas as pd
 import torch.distributed.checkpoint as dcp
 from omegaconf import DictConfig
-from torch import multiprocessing as mp
-from torch.distributed.checkpoint import FileSystemWriter
-
-import os
 import torch
 import torch.distributed as dist
 import torch.utils.data
@@ -26,12 +20,8 @@ from torch.distributed.fsdp.wrap import transformer_auto_wrap_policy
 from torch.distributed.fsdp.fully_sharded_data_parallel import StateDictType
 from transformers.models.llama.modeling_llama import LlamaDecoderLayer
 
-from s3torchbenchmarking.benchmark_utils import (
-    build_random_suffix,
-    build_checkpoint_uri,
-)
+from s3torchbenchmarking.dcp_common import setup, run_benchmark_common, get_writer
 from s3torchbenchmarking.models import get_benchmark_model
-from s3torchconnector.dcp import S3StorageWriter
 
 Timestamps = Tuple[float, float]
 logger = logging.getLogger(__name__)
@@ -41,77 +31,10 @@ logger = logging.getLogger(__name__)
 @hydra.main(version_base=None)
 def run_benchmark(cfg: DictConfig) -> dict:
     """DCP benchmarks entry point."""
-    manager = mp.Manager()
-    corrected_save_timestamps: Queue[Timestamps] = manager.Queue()
-    processing_timestamps: List[Timestamps] = []
-
-    suffix = build_random_suffix()
-
-    for epoch in range(cfg.epochs):
-        logger.info("Executing epoch #%i / %i...", epoch + 1, cfg.epochs)
-        begin_mp = perf_counter()
-        mp.spawn(
-            _run,
-            (cfg, suffix, corrected_save_timestamps),
-            nprocs=cfg.world_size,
-            join=True,
-        )
-        end_mp = perf_counter()
-        processing_timestamps.append((begin_mp, end_mp))
-
-    # Dump the multiprocessing Queue's content into a list.
-    collector: List[Timestamps] = []
-    while not corrected_save_timestamps.empty():
-        collector.append(corrected_save_timestamps.get())
-
-    # Collect all data in Pandas DataFrame, and dump them (through `describe()`) in a Python dict.
-    cst = pd.DataFrame(collector, columns=["begin", "end", "size"])
-    pt = pd.DataFrame(processing_timestamps, columns=["begin", "end"])
-
-    corrected_save_durations_s = cst["end"] - cst["begin"]
-    processing_durations_s = pt["end"] - pt["begin"]
-    # Expectation that all workers started at the same time. So we are taking the longest time
-    # to get average throughput for the entire run.
-    throughput_mibs = cst["size"] / corrected_save_durations_s.max()
-
-    metrics = {
-        "throughput_mibs": throughput_mibs.dropna().to_list(),
-        "corrected_save_durations_s": corrected_save_durations_s.dropna().to_list(),
-        "processing_durations_s": processing_durations_s.dropna().to_list(),
-    }
-    return {"metrics": metrics}
+    return run_benchmark_common(cfg, run_fsdp, (cfg,))
 
 
-def get_writer(cfg: DictConfig, suffix) -> FileSystemWriter:
-    """Instantiate a checkpoint writer based on the input config."""
-
-    if cfg.checkpoint.storage == "disk":
-        local_path = Path(cfg.path) / suffix
-        logger.info("Saving checkpoint to %s (disk)...", local_path)
-        sync_files = (
-            True
-            if cfg.checkpoint.sync_files is not None and cfg.checkpoint.sync_files
-            else False
-        )
-        return dcp.FileSystemWriter(
-            local_path, sync_files=sync_files, thread_count=cfg.thread_count
-        )
-    elif cfg.checkpoint.storage == "s3":
-        uri = build_checkpoint_uri(cfg.s3.uri, suffix)
-        logger.info("Saving checkpoint to %s (S3)...", uri)
-        return S3StorageWriter(cfg.s3.region, uri, thread_count=cfg.thread_count)
-    else:
-        raise ValueError(f"Storage writer {cfg.checkpoint.storage} not supported")
-
-
-def setup(backend: str, world_size: int, rank: int) -> None:
-    os.environ["MASTER_ADDR"] = "localhost"
-    os.environ["MASTER_PORT"] = "12355"
-    os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "expandable_segments:True"
-    dist.init_process_group(backend, world_size=world_size, rank=rank)
-
-
-def _run(
+def run_fsdp(
     rank: int,  # needs to be passed first (provided by `multiprocessing.spawn` automatically)
     cfg: DictConfig,
     suffix,
